Fixed LineChart Consol Warning
